+++ CHANGELOG.md
@@ -8,9 +8,14 @@ This project adheres to [Semantic Versioning 2.0.0](https://semver.org/spec/v2.0
 
 ### Added
 
+- Telegram Bot notification support:
+  - enabled telegram notification type
+  - user registration flow to custom bot (bot config admin-side, user registration in settings page)
+- frontend direct acces to specific event details (Task detail open modal from link)
+
 ### Changed
 
-- UX improvements: condensed headers/navigation and space optimization
+- UI improvements: condensed headers/navigation and space optimization
 
 ### Deprecated